pyproject.toml
[project]
name = "riskcalculator"
version = "0.1.1"
description = ""
authors = [
    {name = "Daniel"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "schwab-py (>=1.5.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "cufflinks (>=0.17.3,<0.18.0)",
    "numpy (==1.26.4)",
    "nbformat (>=5.10.4,<6.0.0)",
    "mplfinance (>=0.12.10b0,<0.13.0)",
    "xlsxwriter (>=3.2.5,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "alpha-vantage (>=3.0.0,<4.0.0)",
    "lxml (>=6.0.1,<7.0.0)",
    "edgartools (>=4.8.2,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"

